% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{respects_order}
\alias{respects_order}
\title{Helper function that checks a conditioning index against a particular ordering.}
\usage{
respects_order(index_given, ordering)
}
\arguments{
\item{index_given}{}
}
\description{
Helper function that checks a conditioning index against a particular ordering.
}
\keyword{internal}
